+++ packages/axios-extension/src/abap/types/adt-types.ts
@@ -133,3 +133,11 @@ export interface ArchiveFileNode {
 export type ArchiveFileNodeType = 'file' | 'folder';
 
 export type ArchiveFileContentType<T> = T extends 'file' ? string : T extends 'folder' ? ArchiveFileNode[] : never;
+
+export type BusinessObject = {
+    name: string;
+    uri: string;
+    description?: string;
+};
+
+export type PublishResponse = { SEVERITY: string; SHORT_TEXT: string; LONG_TEXT: string };