+++ packages/odata-service-inquirer/test/unit/index-api.test.ts
@@ -1,9 +1,8 @@
-import { initI18nOdataServiceInquirer } from '../../src/i18n';
-import * as prompts from '../../src/prompts';
-import { PromptState } from '../../src/utils';
-import { getPrompts } from '../../src/index';
-import LoggerHelper from '../../src/prompts/logger-helper';
 import { ErrorHandler } from '../../src/error-handler/error-handler';
+import { getPrompts } from '../../src/index';
+import * as prompts from '../../src/prompts';
+import LoggerHelper from '../../src/prompts/logger-helper';
+import { PromptState } from '../../src/utils';
 
 jest.mock('../../src/prompts', () => ({
     __esModule: true, // Workaround to for spyOn TypeError: Jest cannot redefine property
@@ -11,9 +10,8 @@ jest.mock('../../src/prompts', () => ({
 }));
 
 describe('API tests', () => {
-    beforeAll(async () => {
-        // Wait for i18n to bootstrap so we can test localised strings
-        await initI18nOdataServiceInquirer();
+    beforeEach(() => {
+        jest.restoreAllMocks();
     });
 
     test('getPrompts', async () => {
@@ -36,5 +34,98 @@ describe('API tests', () => {
         expect(LoggerHelper.logger).toBeDefined();
         expect(ErrorHandler.guidedAnswersEnabled).toBe(true);
         expect(ErrorHandler.logger).toBeDefined();
+    });
+
+    test('getPrompts, i18n is loaded', async () => {
+        const { prompts: questions } = await getPrompts(undefined, undefined, true, undefined, true);
+
+        expect(questions).toMatchInlineSnapshot(`
+            [
+              {
+                "additionalMessages": [Function],
+                "choices": [
+                  {
+                    "name": "Connect to a SAP System",
+                    "value": "sapSystem",
+                  },
+                  {
+                    "name": "Connect to an OData Service Url",
+                    "value": "odataServiceUrl",
+                  },
+                  {
+                    "name": "Connect to SAP Business Accelerator Hub",
+                    "value": "businessHub",
+                  },
+                  {
+                    "name": "Use a Local CAP Project",
+                    "value": "capProject",
+                  },
+                  {
+                    "name": "Upload a Metadata File",
+                    "value": "metadataFile",
+                  },
+                ],
+                "default": -1,
+                "guiOptions": {
+                  "breadcrumb": true,
+                },
+                "message": "Data source",
+                "name": "datasourceType",
+                "type": "list",
+              },
+              {
+                "guiOptions": {
+                  "breadcrumb": true,
+                  "mandatory": true,
+                },
+                "guiType": "file-browser",
+                "message": "Metadata file path",
+                "name": "metadataFilePath",
+                "type": "input",
+                "validate": [Function],
+                "when": [Function],
+              },
+              {
+                "choices": [Function],
+                "default": [Function],
+                "guiOptions": {
+                  "applyDefaultWhenDirty": true,
+                  "breadcrumb": "CAP Project",
+                  "mandatory": true,
+                },
+                "message": "Choose your CAP project",
+                "name": "capProject",
+                "type": "list",
+                "when": [Function],
+              },
+              {
+                "default": [Function],
+                "guiOptions": {
+                  "breadcrumb": "CAP Project",
+                  "mandatory": true,
+                },
+                "guiType": "folder-browser",
+                "message": "CAP project folder path",
+                "name": "capProjectPath",
+                "type": "input",
+                "validate": [Function],
+                "when": [Function],
+              },
+              {
+                "choices": [Function],
+                "default": [Function],
+                "guiOptions": {
+                  "applyDefaultWhenDirty": true,
+                  "breadcrumb": true,
+                  "mandatory": true,
+                },
+                "message": "OData service",
+                "name": "capService",
+                "type": "list",
+                "validate": [Function],
+                "when": [Function],
+              },
+            ]
+        `);
     });
 });